+++ changelog.md
@@ -1,6 +1,18 @@
 # Changelog
 
-**./test-parts/part-c-internal_scripts/scripts/s-local_valid.js**
-* Removed functions:
-	* handleIpSixFunction
-	* handleMacAddressFunction
+**./app/common-errors.js**
+* Restructured functions to write text on single line:
+	* writeKeyNotFoundErrorString
+	* writeRegisterPrefixIndexError
+* Restructured functions to write a single variable:
+	* writeRemoteIoPropertyErrorString
+	* writeRemoteIoPropertySupportString
+	* writeRemoteIoPropertyConstructString
+	* writeConnectDevicePropertyErrorString
+	* writeSetDeviceOutputWrongError
+	* writeUnexpectedTokenErrorString
+* Renamed result variables in these functions:
+	* 'writeUnexpectedTokenErrorStringNull' to 'nullRes'
+	* 'writeUnexpectedTokenErrorStringType' to 'typeRes'
+	* 'writeTestString' to 'testValue'
+* Replaced `exports` with `module.exports`